+++ src/api/postToken.ts
@@ -1,0 +1,8 @@
+// import axios from "axios";
+
+// export async function postToken(token: string) {
+//   const data = await axios.get(`43.201.69.115:8080/notifications`, token);
+
+//   console.log(data);
+//   return data;
+// }
