+++ src/firebase-messaging-sw.js
@@ -1,0 +1,41 @@
+import { initializeApp } from "firebase/app";
+import { getMessaging, getToken, onMessage } from "firebase/messaging";
+
+const firebaseConfig = {
+  apiKey: import.meta.env.VITE_APP_API_KEY,
+  authDomain: import.meta.env.VITE_APP_AUTH_DOMAIN,
+  projectId: import.meta.env.VITE_APP_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_APP_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_APP_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_APP_APP_ID,
+  measurementId: import.meta.env.VITE_APP_MEASUREMENT_ID,
+};
+
+const app = initializeApp(firebaseConfig);
+const messaging = getMessaging(app);
+
+async function requestPermission() {
+  console.log("권한 요청 중...");
+
+  const permission = await Notification.requestPermission();
+  if (permission === "denied") {
+    console.log("알림 권한 허용 안됨");
+    return;
+  }
+
+  console.log("알림 권한이 허용됨");
+
+  const token = await getToken(messaging, {
+    vapidKey: import.meta.env.VITE_APP_VAPID_KEY,
+  });
+
+  if (token) console.log("token: ", token);
+  else console.log("Can not get Token");
+
+  onMessage(messaging, (payload) => {
+    console.log("메시지가 도착했습니다.", payload);
+    // ...
+  });
+}
+
+requestPermission();
